requirements.txt.txt
networkx==2.5.1
osmnx==1.1.0
pickleshare==0.7.5
urllib3==1.25.8
python-csv==0.0.13
scikit-learn==0.24.2
staticmap==0.5.5
python-telegram-bot==13.5
python-dateutil==2.8.1